+++ src/components/ui/Header/HeaderDropdown.tsx
@@ -1,10 +1,11 @@
-import { Menu, Transition } from '@headlessui/react'
-import DownArrow from './svg/DownArrow';
+import { Menu, Transition } from "@headlessui/react";
+import DownArrow from "../svg/DownArrow";
 
-const HeaderDropdown = (
-  { disconnect }:
-    { disconnect: () => void }
-) => {
+interface HeaderDropdownProps {
+  disconnect: () => void;
+}
+
+const HeaderDropdown = ({ disconnect }: HeaderDropdownProps) => {
   return (
     <div className="flex items-center justify-center">
       <div className="relative inline-block text-left">
@@ -12,7 +13,7 @@ const HeaderDropdown = (
           {({ open }) => (
             <>
               <Menu.Button className="text-sm font-medium text-white transition duration-150 ease-in-out hover:text-gray-500 focus:outline-none focus:border-blue-300 focus:shadow-outline-blue">
-                <DownArrow/>
+                <DownArrow />
               </Menu.Button>
 
               <Transition
@@ -33,10 +34,9 @@ const HeaderDropdown = (
                       {({ active }) => (
                         <button
                           onClick={disconnect}
-                          className={`${active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700"
-                            } flex justify-between w-full px-4 py-2 text-sm leading-5 text-left`}
+                          className={`${
+                            active ? "bg-gray-100 text-gray-900" : "text-gray-700"
+                          } flex justify-between w-full px-4 py-2 text-sm leading-5 text-left`}
                         >
                           Disconnect
                         </button>
@@ -51,7 +51,6 @@ const HeaderDropdown = (
       </div>
     </div>
   );
+};
 
-}
-
-export default HeaderDropdown
+export default HeaderDropdown;
